+++ src/00_core.lua
@@ -24,7 +24,7 @@
 
 _G.Ess = _G.Ess or {}
 local Ess = _G.Ess
-Ess.VERSION = "0.2.0"
+Ess.VERSION = "0.2.1"
 
 Ess.Safe = Ess.Safe or {}
 Ess.Table = Ess.Table or {}